+++ front-end/src/app/components/hint/hint.component.ts
@@ -9,8 +9,7 @@ import { MqttService } from "ngx-mqtt";
 export class HintComponent {
   hint: string;
 
-  constructor(private mqttService: MqttService) {
-  }
+  constructor(private mqttService: MqttService) {}
 
   public unsafePublish(topic: string, message: string): void {
     this.mqttService.unsafePublish(topic, message, { qos: 1, retain: true });
