--- front-end/src/app/components/interfaces/device.ts
@@ -1,6 +1,0 @@
-export interface Device {
-  id: string;
-  description: string;
-  status: string[];
-  connection: boolean;
-}
